feat(vm): cell value e channel buf da int → int64_t
vm_types.h: Var.value e Channel.buf passano da int* a int64_t*. Aggiunto
#include <stdint.h>.

Propagazione:
- vm_helpers.h: resolve_atom/resolve_expr/resolve_value ritornano int64_t;
  strtol → strtoll; tutti sizeof(int) per value/buf → sizeof(int64_t).
- vm_ops.h: tmp/val/popped/marker/n locali letti da value/buf passano a
  int64_t. Print "%d" → "%lld" con cast (long long). strtoul → strtoull.
  eval_cond signature int → int64_t (parametri lval, rval).
- vm_invert.h: calloc(1, sizeof(int)) per value → sizeof(int64_t).
- vm_debug.h, Janus.c: arr pointer e printf "%d" su valori → int64_t/%lld.

Abilita cell con valori > INT_MAX (unsigned 32-bit interpretation come
4294967295). Necessario per Mnemo emit di `%u` runtime su negative
(`if cell < 0 then cell += 2^32`).

Limitazione residua: `__mn_divmod_nonneg` è O(n) per sottrazione ripetuta.
Per n=2^32 → catastrofico (~ 4G iterazioni). Lato Mnemo il fix `%u`
runtime negative resta disabilitato finché divmod non ha algoritmo
sub-lineare (doubling-based).

+++ src/vm/Janus.c
@@ -565,13 +565,13 @@ void vm_dump(VM *vm)
             Var *v = f->vars[j]; if (!v) continue;
             vm_printf("%s: ", v->name);
             if (v->T == TYPE_INT) {
-                vm_printf("%d", *(v->value));
+                vm_printf("%lld", (long long)*(v->value));
             } else {
                 vm_printf("[");
                 size_t n = (v->T == TYPE_STACK) ? v->stack_len : v->channel->buf_len;
-                int *arr = (v->T == TYPE_STACK) ? v->value : v->channel->buf;
+                int64_t *arr = (v->T == TYPE_STACK) ? v->value : v->channel->buf;
                 for (size_t k = 0; k < n; k++) {
-                    vm_printf("%d", arr[k]);
+                    vm_printf("%lld", (long long)arr[k]);
                     if (k + 1 < n) vm_printf(", ");
                 }
                 vm_printf("]");
